back

+++ src/main/java/com/swpproject/application/model/SlotExcercise.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.util.List;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
@@ -15,12 +17,11 @@ public class SlotExcercise {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String day;
     private String start_hour;
     private String end_hour;
-
     private boolean is_pending;
     private String attendant_status;
